+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - `database::set_errmsg` wrapping `sqlite3_set_errmsg`
 - `keywords` class providing safe, STL-like interface to `sqlite3_keyword_count`, `sqlite3_keyword_name` and `sqlite3_keyword_check`
 - `error::offset` wrapping `sqlite3_error_offset`
+- `thinsqlitepp::status` wrapping `sqlite3_status` or `sqlite3_status64`
 - clang-cl is now supported on Windows
 
 ### Fixed